0906 close #28 使之可以對應linux

+++ app/autoit/ImageMagickHelper.js
@@ -6,7 +6,8 @@ let ImageMagickHelper = {
     //imagemagick: null,
     fs: null,
     path: null,
-    exec: null
+    exec: null,
+    icoToPng: null
   },
   imagemagickPath: null,
   init: function () {
@@ -19,6 +20,7 @@ let ImageMagickHelper = {
     this.lib.exec = require('child_process').exec
     
     this.lib.imageSize = require('image-size')
+    this.lib.icoToPng = require('ico-to-png')
     this.imagemagickPath = this.lib.path.resolve(__dirname, '../win32-helpers/imagemagick/convert.exe')
     //this.lib.imagemagick = require('imagemagick')
     //this.lib.gm = require('gm').subClass({imageMagick: false})
@@ -71,16 +73,25 @@ let ImageMagickHelper = {
       return callback(pngPath)
     }
 
-    console.log(['Start convert: ', icoPath, pngPath])
-    
-    let command = `"${this.imagemagickPath}" convert "${icoPath}" "${pngPath}"`
-    console.log(command)
-    this.lib.exec(command, (err) => {
-      if (err) {
-        console.error(err)
-      }
-      callback(pngPath)
-    })
+    if (process.platform === 'win32') {
+      console.log(['Start convert: ', icoPath, pngPath])
+
+      let command = `"${this.imagemagickPath}" convert "${icoPath}" "${pngPath}"`
+      console.log(command)
+      this.lib.exec(command, (err) => {
+        if (err) {
+          console.error(err)
+        }
+        callback(pngPath)
+      })
+    }
+    else {
+      this.lib.icoToPng(source).then((pngBuffer) => {
+        this.lib.fs.writeFile(pngPath, pngBuffer, () => {
+          callback(pngPath)
+        })
+      })
+    }
     
     return this
     /*
@@ -112,12 +123,21 @@ let ImageMagickHelper = {
       return this
     }
     
-    return this.icoToPng(icoPath, (pngPath) => {
-      console.log(pngPath)
-      let pngBuffer = this.lib.fs.readFileSync(pngPath)
-      callback('data:image/png;base64,' + new Buffer(pngBuffer).toString('base64'))
-      this.lib.fs.unlinkSync(pngPath)
-    })
+    if (process.platform === 'win32') {
+      return this.icoToPng(icoPath, (pngPath) => {
+        console.log(pngPath)
+        let pngBuffer = this.lib.fs.readFileSync(pngPath)
+        callback('data:image/png;base64,' + new Buffer(pngBuffer).toString('base64'))
+        this.lib.fs.unlinkSync(pngPath)
+      })
+    }
+    else {
+      let source = this.lib.fs.readFileSync(icoPath)
+      this.lib.icoToPng(source).then((pngBuffer) => {
+        callback('data:image/png;base64,' + new Buffer(pngBuffer).toString('base64'))
+      })
+      return this
+    }
   }
 }
